+++ resources/hpxml-measures/BuildResidentialHPXML/measure.rb
@@ -1382,6 +1382,76 @@ class BuildResidentialHPXML < OpenStudio::Measure::ModelMeasure
     arg.setUnits('W')
     args << arg
 
+    perf_data_capacity_type_choices = OpenStudio::StringVector.new
+    perf_data_capacity_type_choices << 'Absolute capacities'
+    # perf_data_capacity_type_choices << 'Normalized capacity fractions'
+
+    arg = OpenStudio::Measure::OSArgument.makeChoiceArgument('hvac_perf_data_capacity_type', perf_data_capacity_type_choices, false)
+    arg.setDisplayName('HVAC Detailed Performance Data: Capacity Type')
+    arg.setDescription('Type of capacity values for detailed performance data if available. Applies only to variable-speed air-source HVAC systems (central air conditioners, mini-split air conditioners, air-to-air heat pumps, and mini-split heat pumps).')
+    arg.setUnits('Absolute capacities')
+    args << arg
+
+    arg = OpenStudio::Measure::OSArgument.makeStringArgument('hvac_perf_data_heating_outdoor_temperatures', false)
+    arg.setDisplayName('HVAC Detailed Performance Data: Heating Outdoor Temperatures')
+    arg.setDescription('Outdoor temperatures of heating detailed performance data if available. Applies only to variable-speed air-source HVAC systems (central air conditioners, mini-split air conditioners, air-to-air heat pumps, and mini-split heat pumps). One of the outdoor temperatures must be 47 deg-F. At least two performance data points are required using a comma-separated list.')
+    arg.setUnits('deg-F')
+    args << arg
+
+    arg = OpenStudio::Measure::OSArgument.makeStringArgument('hvac_perf_data_heating_min_speed_capacities', false)
+    arg.setDisplayName('HVAC Detailed Performance Data: Heating Minimum Speed Capacities')
+    arg.setDescription('Minimum speed capacities of heating detailed performance data if available. Applies only to variable-speed air-source HVAC systems (central air conditioners, mini-split air conditioners, air-to-air heat pumps, and mini-split heat pumps). At least two performance data points are required using a comma-separated list.')
+    arg.setUnits('Btu/hr or Frac')
+    args << arg
+
+    arg = OpenStudio::Measure::OSArgument.makeStringArgument('hvac_perf_data_heating_max_speed_capacities', false)
+    arg.setDisplayName('HVAC Detailed Performance Data: Heating Maximum Speed Capacities')
+    arg.setDescription('Maximum speed capacities of heating detailed performance data if available. Applies only to variable-speed air-source HVAC systems (central air conditioners, mini-split air conditioners, air-to-air heat pumps, and mini-split heat pumps). At least two performance data points are required using a comma-separated list.')
+    arg.setUnits('Btu/hr or Frac')
+    args << arg
+
+    arg = OpenStudio::Measure::OSArgument.makeStringArgument('hvac_perf_data_heating_min_speed_cops', false)
+    arg.setDisplayName('HVAC Detailed Performance Data: Heating Minimum Speed COPs')
+    arg.setDescription('Minimum speed efficiency COP values of heating detailed performance data if available. Applies only to variable-speed air-source HVAC systems (central air conditioners, mini-split air conditioners, air-to-air heat pumps, and mini-split heat pumps). At least two performance data points are required using a comma-separated list.')
+    arg.setUnits('W/W')
+    args << arg
+
+    arg = OpenStudio::Measure::OSArgument.makeStringArgument('hvac_perf_data_heating_max_speed_cops', false)
+    arg.setDisplayName('HVAC Detailed Performance Data: Heating Maximum Speed COPs')
+    arg.setDescription('Maximum speed efficiency COP values of heating detailed performance data if available. Applies only to variable-speed air-source HVAC systems (central air conditioners, mini-split air conditioners, air-to-air heat pumps, and mini-split heat pumps). At least two performance data points are required using a comma-separated list.')
+    arg.setUnits('W/W')
+    args << arg
+
+    arg = OpenStudio::Measure::OSArgument.makeStringArgument('hvac_perf_data_cooling_outdoor_temperatures', false)
+    arg.setDisplayName('HVAC Detailed Performance Data: Cooling Outdoor Temperatures')
+    arg.setDescription('Outdoor temperatures of cooling detailed performance data if available. Applies only to variable-speed air-source HVAC systems (central air conditioners, mini-split air conditioners, air-to-air heat pumps, and mini-split heat pumps). One of the outdoor temperatures must be 95 deg-F. At least two performance data points are required using a comma-separated list.')
+    arg.setUnits('deg-F')
+    args << arg
+
+    arg = OpenStudio::Measure::OSArgument.makeStringArgument('hvac_perf_data_cooling_min_speed_capacities', false)
+    arg.setDisplayName('HVAC Detailed Performance Data: Cooling Minimum Speed Capacities')
+    arg.setDescription('Minimum speed capacities of cooling detailed performance data if available. Applies only to variable-speed air-source HVAC systems (central air conditioners, mini-split air conditioners, air-to-air heat pumps, and mini-split heat pumps). At least two performance data points are required using a comma-separated list.')
+    arg.setUnits('Btu/hr or Frac')
+    args << arg
+
+    arg = OpenStudio::Measure::OSArgument.makeStringArgument('hvac_perf_data_cooling_max_speed_capacities', false)
+    arg.setDisplayName('HVAC Detailed Performance Data: Cooling Maximum Speed Capacities')
+    arg.setDescription('Maximum speed capacities of cooling detailed performance data if available. Applies only to variable-speed air-source HVAC systems (central air conditioners, mini-split air conditioners, air-to-air heat pumps, and mini-split heat pumps). At least two performance data points are required using a comma-separated list.')
+    arg.setUnits('Btu/hr or Frac')
+    args << arg
+
+    arg = OpenStudio::Measure::OSArgument.makeStringArgument('hvac_perf_data_cooling_min_speed_cops', false)
+    arg.setDisplayName('HVAC Detailed Performance Data: Cooling Minimum Speed COPs')
+    arg.setDescription('Minimum speed efficiency COP values of cooling detailed performance data if available. Applies only to variable-speed air-source HVAC systems (central air conditioners, mini-split air conditioners, air-to-air heat pumps, and mini-split heat pumps). At least two performance data points are required using a comma-separated list.')
+    arg.setUnits('W/W')
+    args << arg
+
+    arg = OpenStudio::Measure::OSArgument.makeStringArgument('hvac_perf_data_cooling_max_speed_cops', false)
+    arg.setDisplayName('HVAC Detailed Performance Data: Cooling Maximum Speed COPs')
+    arg.setDescription('Maximum speed efficiency COP values of cooling detailed performance data if available. Applies only to variable-speed air-source HVAC systems (central air conditioners, mini-split air conditioners, air-to-air heat pumps, and mini-split heat pumps). At least two performance data points are required using a comma-separated list.')
+    arg.setUnits('W/W')
+    args << arg
+
     geothermal_loop_configuration_choices = OpenStudio::StringVector.new
     geothermal_loop_configuration_choices << 'none'
     # geothermal_loop_configuration_choices << HPXML::GeothermalLoopLoopConfigurationDiagonal
@@ -3395,6 +3465,44 @@ class BuildResidentialHPXML < OpenStudio::Measure::ModelMeasure
     error = args[:rim_joist_assembly_r].is_initialized && !args[:geometry_rim_joist_height].is_initialized
     errors << 'Specified a rim joist assembly R-value but no rim joist height.' if error
 
+    hvac_perf_data_heating_args_initialized = [args[:hvac_perf_data_heating_outdoor_temperatures].is_initialized,
+                                               args[:hvac_perf_data_heating_min_speed_capacities].is_initialized,
+                                               args[:hvac_perf_data_heating_max_speed_capacities].is_initialized,
+                                               args[:hvac_perf_data_heating_min_speed_cops].is_initialized,
+                                               args[:hvac_perf_data_heating_max_speed_cops].is_initialized]
+    error = (hvac_perf_data_heating_args_initialized.uniq.size != 1)
+    errors << 'Did not specify all required heating detailed performance data arguments.' if error
+
+    if hvac_perf_data_heating_args_initialized.uniq.size == 1 && hvac_perf_data_heating_args_initialized.uniq[0]
+      heating_data_points_lengths = [args[:hvac_perf_data_heating_outdoor_temperatures].get.count(','),
+                                     args[:hvac_perf_data_heating_min_speed_capacities].get.count(','),
+                                     args[:hvac_perf_data_heating_max_speed_capacities].get.count(','),
+                                     args[:hvac_perf_data_heating_min_speed_cops].get.count(','),
+                                     args[:hvac_perf_data_heating_max_speed_cops].get.count(',')]
+
+      error = (heating_data_points_lengths.uniq.size != 1)
+      errors << 'One or more detailed heating performance data arguments does not have enough comma-separated elements specified.' if error
+    end
+
+    hvac_perf_data_cooling_args_initialized = [args[:hvac_perf_data_cooling_outdoor_temperatures].is_initialized,
+                                               args[:hvac_perf_data_cooling_min_speed_capacities].is_initialized,
+                                               args[:hvac_perf_data_cooling_max_speed_capacities].is_initialized,
+                                               args[:hvac_perf_data_cooling_min_speed_cops].is_initialized,
+                                               args[:hvac_perf_data_cooling_max_speed_cops].is_initialized]
+    error = (hvac_perf_data_cooling_args_initialized.uniq.size != 1)
+    errors << 'Did not specify all required cooling detailed performance data arguments.' if error
+
+    if hvac_perf_data_cooling_args_initialized.uniq.size == 1 && hvac_perf_data_cooling_args_initialized.uniq[0]
+      cooling_data_points_lengths = [args[:hvac_perf_data_cooling_outdoor_temperatures].get.count(','),
+                                     args[:hvac_perf_data_cooling_min_speed_capacities].get.count(','),
+                                     args[:hvac_perf_data_cooling_max_speed_capacities].get.count(','),
+                                     args[:hvac_perf_data_cooling_min_speed_cops].get.count(','),
+                                     args[:hvac_perf_data_cooling_max_speed_cops].get.count(',')]
+
+      error = (cooling_data_points_lengths.uniq.size != 1)
+      errors << 'One or more detailed cooling performance data arguments does not have enough comma-separated elements specified.' if error
+    end
+
     emissions_args_initialized = [args[:emissions_scenario_names].is_initialized,
                                   args[:emissions_types].is_initialized,
                                   args[:emissions_electricity_units].is_initialized,
@@ -5106,6 +5214,43 @@ class HPXMLFile
                                    integrated_heating_system_capacity: integrated_heating_system_capacity,
                                    integrated_heating_system_efficiency_percent: integrated_heating_system_efficiency_percent,
                                    integrated_heating_system_fraction_heat_load_served: integrated_heating_system_fraction_heat_load_served)
+
+    if args[:hvac_perf_data_capacity_type].is_initialized && [HPXML::HVACTypeCentralAirConditioner, HPXML::HVACTypeMiniSplitAirConditioner].include?(cooling_system_type) && compressor_type == HPXML::HVACCompressorTypeVariableSpeed
+      hvac_perf_data_capacity_type = args[:hvac_perf_data_capacity_type].get
+      hvac_perf_data_cooling_outdoor_temperatures = args[:hvac_perf_data_cooling_outdoor_temperatures].get.split(',').map(&:strip)
+      hvac_perf_data_cooling_min_speed_capacities = args[:hvac_perf_data_cooling_min_speed_capacities].get.split(',').map(&:strip)
+      hvac_perf_data_cooling_max_speed_capacities = args[:hvac_perf_data_cooling_max_speed_capacities].get.split(',').map(&:strip)
+      hvac_perf_data_cooling_min_speed_cops = args[:hvac_perf_data_cooling_min_speed_cops].get.split(',').map(&:strip)
+      hvac_perf_data_cooling_max_speed_cops = args[:hvac_perf_data_cooling_max_speed_cops].get.split(',').map(&:strip)
+
+      clg_perf_data = hpxml_bldg.cooling_systems[0].cooling_detailed_performance_data
+      cooling_perf_data_data_points = hvac_perf_data_cooling_outdoor_temperatures.zip(hvac_perf_data_cooling_min_speed_capacities,
+                                                                                      hvac_perf_data_cooling_max_speed_capacities,
+                                                                                      hvac_perf_data_cooling_min_speed_cops,
+                                                                                      hvac_perf_data_cooling_max_speed_cops)
+      cooling_perf_data_data_points.each do |cooling_perf_data_data_point|
+        outdoor_temperature, min_speed_cap_or_frac, max_speed_cap_or_frac, min_speed_cop, max_speed_cop = cooling_perf_data_data_point
+
+        if hvac_perf_data_capacity_type == 'Absolute capacities'
+          min_speed_capacity = min_speed_cap_or_frac
+          max_speed_capacity = max_speed_cap_or_frac
+        elsif hvac_perf_data_capacity_type == 'Normalized capacity fractions'
+          min_speed_capacity_fraction_of_nominal = min_speed_cap_or_frac
+          max_speed_capacity_fraction_of_nominal = max_speed_cap_or_frac
+        end
+
+        clg_perf_data.add(outdoor_temperature: outdoor_temperature,
+                          capacity: min_speed_capacity,
+                          capacity_fraction_of_nominal: min_speed_capacity_fraction_of_nominal,
+                          capacity_description: HPXML::CapacityDescriptionMinimum,
+                          efficiency_cop: min_speed_cop)
+        clg_perf_data.add(outdoor_temperature: outdoor_temperature,
+                          capacity: max_speed_capacity,
+                          capacity_fraction_of_nominal: max_speed_capacity_fraction_of_nominal,
+                          capacity_description: HPXML::CapacityDescriptionMaximum,
+                          efficiency_cop: max_speed_cop)
+      end
+    end
   end
 
   def self.set_heat_pumps(hpxml_bldg, args)
@@ -5250,6 +5395,80 @@ class HPXMLFile
                               crankcase_heater_watts: heat_pump_crankcase_heater_watts,
                               primary_heating_system: primary_heating_system,
                               primary_cooling_system: primary_cooling_system)
+
+    if args[:hvac_perf_data_capacity_type].is_initialized && [HPXML::HVACTypeHeatPumpAirToAir, HPXML::HVACTypeHeatPumpMiniSplit].include?(heat_pump_type) && compressor_type == HPXML::HVACCompressorTypeVariableSpeed
+      hvac_perf_data_capacity_type = args[:hvac_perf_data_capacity_type].get
+      hvac_perf_data_heating_outdoor_temperatures = args[:hvac_perf_data_heating_outdoor_temperatures].get.split(',').map(&:strip)
+      hvac_perf_data_heating_min_speed_capacities = args[:hvac_perf_data_heating_min_speed_capacities].get.split(',').map(&:strip)
+      hvac_perf_data_heating_max_speed_capacities = args[:hvac_perf_data_heating_max_speed_capacities].get.split(',').map(&:strip)
+      hvac_perf_data_heating_min_speed_cops = args[:hvac_perf_data_heating_min_speed_cops].get.split(',').map(&:strip)
+      hvac_perf_data_heating_max_speed_cops = args[:hvac_perf_data_heating_max_speed_cops].get.split(',').map(&:strip)
+
+      htg_perf_data = hpxml_bldg.heat_pumps[0].heating_detailed_performance_data
+      heating_perf_data_data_points = hvac_perf_data_heating_outdoor_temperatures.zip(hvac_perf_data_heating_min_speed_capacities,
+                                                                                      hvac_perf_data_heating_max_speed_capacities,
+                                                                                      hvac_perf_data_heating_min_speed_cops,
+                                                                                      hvac_perf_data_heating_max_speed_cops)
+      heating_perf_data_data_points.each do |heating_perf_data_data_point|
+        outdoor_temperature, min_speed_cap_or_frac, max_speed_cap_or_frac, min_speed_cop, max_speed_cop = heating_perf_data_data_point
+
+        if hvac_perf_data_capacity_type == 'Absolute capacities'
+          min_speed_capacity = min_speed_cap_or_frac
+          max_speed_capacity = max_speed_cap_or_frac
+        elsif hvac_perf_data_capacity_type == 'Normalized capacity fractions'
+          min_speed_capacity_fraction_of_nominal = min_speed_cap_or_frac
+          max_speed_capacity_fraction_of_nominal = max_speed_cap_or_frac
+        end
+
+        htg_perf_data.add(outdoor_temperature: outdoor_temperature,
+                          capacity: min_speed_capacity,
+                          capacity_fraction_of_nominal: min_speed_capacity_fraction_of_nominal,
+                          capacity_description: HPXML::CapacityDescriptionMinimum,
+                          efficiency_cop: min_speed_cop)
+        htg_perf_data.add(outdoor_temperature: outdoor_temperature,
+                          capacity: max_speed_capacity,
+                          capacity_fraction_of_nominal: max_speed_capacity_fraction_of_nominal,
+                          capacity_description: HPXML::CapacityDescriptionMaximum,
+                          efficiency_cop: max_speed_cop)
+      end
+    end
+
+    if args[:hvac_perf_data_capacity_type].is_initialized
+      hvac_perf_data_capacity_type = args[:hvac_perf_data_capacity_type].get
+      hvac_perf_data_cooling_outdoor_temperatures = args[:hvac_perf_data_cooling_outdoor_temperatures].get.split(',').map(&:strip)
+      hvac_perf_data_cooling_min_speed_capacities = args[:hvac_perf_data_cooling_min_speed_capacities].get.split(',').map(&:strip)
+      hvac_perf_data_cooling_max_speed_capacities = args[:hvac_perf_data_cooling_max_speed_capacities].get.split(',').map(&:strip)
+      hvac_perf_data_cooling_min_speed_cops = args[:hvac_perf_data_cooling_min_speed_cops].get.split(',').map(&:strip)
+      hvac_perf_data_cooling_max_speed_cops = args[:hvac_perf_data_cooling_max_speed_cops].get.split(',').map(&:strip)
+
+      clg_perf_data = hpxml_bldg.heat_pumps[0].cooling_detailed_performance_data
+      cooling_perf_data_data_points = hvac_perf_data_cooling_outdoor_temperatures.zip(hvac_perf_data_cooling_min_speed_capacities,
+                                                                                      hvac_perf_data_cooling_max_speed_capacities,
+                                                                                      hvac_perf_data_cooling_min_speed_cops,
+                                                                                      hvac_perf_data_cooling_max_speed_cops)
+      cooling_perf_data_data_points.each do |cooling_perf_data_data_point|
+        outdoor_temperature, min_speed_cap_or_frac, max_speed_cap_or_frac, min_speed_cop, max_speed_cop = cooling_perf_data_data_point
+
+        if hvac_perf_data_capacity_type == 'Absolute capacities'
+          min_speed_capacity = min_speed_cap_or_frac
+          max_speed_capacity = max_speed_cap_or_frac
+        elsif hvac_perf_data_capacity_type == 'Normalized capacity fractions'
+          min_speed_capacity_fraction_of_nominal = min_speed_cap_or_frac
+          max_speed_capacity_fraction_of_nominal = max_speed_cap_or_frac
+        end
+
+        clg_perf_data.add(outdoor_temperature: outdoor_temperature,
+                          capacity: min_speed_capacity,
+                          capacity_fraction_of_nominal: min_speed_capacity_fraction_of_nominal,
+                          capacity_description: HPXML::CapacityDescriptionMinimum,
+                          efficiency_cop: min_speed_cop)
+        clg_perf_data.add(outdoor_temperature: outdoor_temperature,
+                          capacity: max_speed_capacity,
+                          capacity_fraction_of_nominal: max_speed_capacity_fraction_of_nominal,
+                          capacity_description: HPXML::CapacityDescriptionMaximum,
+                          efficiency_cop: max_speed_cop)
+      end
+    end
   end
 
   def self.set_geothermal_loop(hpxml_bldg, args)