+++ resources/hpxml-measures/BuildResidentialHPXML/tests/test_build_residential_hpxml.rb
@@ -184,6 +184,10 @@ class BuildResidentialHPXMLTest < Minitest::Test
       'error-sfd-with-shared-system.xml' => 'base-sfd.xml',
       'error-rim-joist-height-but-no-assembly-r.xml' => 'base-sfd.xml',
       'error-rim-joist-assembly-r-but-no-height.xml' => 'base-sfd.xml',
+      'error-heating-perf-data-not-all-specified.xml' => 'base-sfd.xml',
+      'error-heating-perf-data-not-all-same-size.xml' => 'base-sfd.xml',
+      'error-cooling-perf-data-not-all-specified.xml' => 'base-sfd.xml',
+      'error-cooling-perf-data-not-all-same-size.xml' => 'base-sfd.xml',
       'error-emissions-args-not-all-specified.xml' => 'base-sfd.xml',
       'error-emissions-args-not-all-same-size.xml' => 'base-sfd.xml',
       'error-emissions-natural-gas-args-not-all-specified.xml' => 'base-sfd.xml',
@@ -244,6 +248,10 @@ class BuildResidentialHPXMLTest < Minitest::Test
       'error-sfd-with-shared-system.xml' => ['Specified a shared system for a single-family detached unit.'],
       'error-rim-joist-height-but-no-assembly-r.xml' => ['Specified a rim joist height but no rim joist assembly R-value.'],
       'error-rim-joist-assembly-r-but-no-height.xml' => ['Specified a rim joist assembly R-value but no rim joist height.'],
+      'error-heating-perf-data-not-all-specified.xml' => ['Did not specify all required heating detailed performance data arguments.'],
+      'error-heating-perf-data-not-all-same-size.xml' => ['One or more detailed heating performance data arguments does not have enough comma-separated elements specified.'],
+      'error-cooling-perf-data-not-all-specified.xml' => ['Did not specify all required cooling detailed performance data arguments.'],
+      'error-cooling-perf-data-not-all-same-size.xml' => ['One or more detailed cooling performance data arguments does not have enough comma-separated elements specified.'],
       'error-emissions-args-not-all-specified.xml' => ['Did not specify all required emissions arguments.'],
       'error-emissions-args-not-all-same-size.xml' => ['One or more emissions arguments does not have enough comma-separated elements specified.'],
       'error-emissions-natural-gas-args-not-all-specified.xml' => ['Did not specify fossil fuel emissions units for natural gas emissions values.'],
@@ -1154,6 +1162,22 @@ class BuildResidentialHPXMLTest < Minitest::Test
       args.delete('rim_joist_assembly_r')
     elsif ['error-rim-joist-assembly-r-but-no-height.xml'].include? hpxml_file
       args.delete('geometry_rim_joist_height')
+    elsif ['error-heating-perf-data-not-all-specified.xml'].include? hpxml_file
+      args['hvac_perf_data_heating_outdoor_temperatures'] = '47.0'
+    elsif ['error-heating-perf-data-not-all-same-size.xml'].include? hpxml_file
+      args['hvac_perf_data_heating_outdoor_temperatures'] = '47.0'
+      args['hvac_perf_data_heating_min_speed_capacities'] = '10000, 4200'
+      args['hvac_perf_data_heating_max_speed_capacities'] = '36000, 24800'
+      args['hvac_perf_data_heating_min_speed_cops'] = '4.73, 1.84'
+      args['hvac_perf_data_heating_max_speed_cops'] = '3.44, 2.66'
+    elsif ['error-cooling-perf-data-not-all-specified.xml'].include? hpxml_file
+      args['hvac_perf_data_cooling_outdoor_temperatures'] = '95.0'
+    elsif ['error-cooling-perf-data-not-all-same-size.xml'].include? hpxml_file
+      args['hvac_perf_data_cooling_outdoor_temperatures'] = '95.0'
+      args['hvac_perf_data_cooling_min_speed_capacities'] = '11700, 13200'
+      args['hvac_perf_data_cooling_max_speed_capacities'] = '36000, 40000'
+      args['hvac_perf_data_cooling_min_speed_cops'] = '4.47, 6.34'
+      args['hvac_perf_data_cooling_max_speed_cops'] = '2.71, 3.53'
     elsif ['error-emissions-args-not-all-specified.xml'].include? hpxml_file
       args['emissions_scenario_names'] = 'Scenario1'
     elsif ['error-emissions-args-not-all-same-size.xml'].include? hpxml_file