+++ resources/hpxml-measures/tasks.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+OpenStudio::Logger.instance.standardOutLogger.setLogLevel(OpenStudio::Fatal)
+
 Dir["#{File.dirname(__FILE__)}/HPXMLtoOpenStudio/resources/*.rb"].each do |resource_file|
   next if resource_file.include? 'minitest_helper.rb'
 
@@ -257,266 +259,6 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     hpxml_bldg.header.manualj_internal_loads_sensible = 4000
     hpxml_bldg.header.manualj_internal_loads_latent = 200
     hpxml_bldg.header.manualj_num_occupants = 5
-  end
-  if ['base-hvac-air-to-air-heat-pump-var-speed-detailed-performance.xml'].include? hpxml_file
-    # YORK HMH7
-    # https://ashp.neep.org/#!/product/64253/7/25000///0
-    clg_perf_data = hpxml_bldg.heat_pumps[0].cooling_detailed_performance_data
-    htg_perf_data = hpxml_bldg.heat_pumps[0].heating_detailed_performance_data
-    clg_perf_data.add(outdoor_temperature: 95.0,
-                      capacity: 11700,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 4.47)
-    clg_perf_data.add(outdoor_temperature: 95.0,
-                      capacity: 36000,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 2.71)
-    clg_perf_data.add(outdoor_temperature: 82.0,
-                      capacity: 13200,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 6.34)
-    clg_perf_data.add(outdoor_temperature: 82.0,
-                      capacity: 40000,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 3.53)
-    htg_perf_data.add(outdoor_temperature: 47.0,
-                      capacity: 10000,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 4.73)
-    htg_perf_data.add(outdoor_temperature: 47.0,
-                      capacity: 36000,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 3.44)
-    htg_perf_data.add(outdoor_temperature: 17.0,
-                      capacity: 4200,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 1.84)
-    htg_perf_data.add(outdoor_temperature: 17.0,
-                      capacity: 24800,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 2.66)
-    htg_perf_data.add(outdoor_temperature: 5.0,
-                      capacity: 1900,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 0.81)
-    htg_perf_data.add(outdoor_temperature: 5.0,
-                      capacity: 19900,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 2.28)
-  end
-  if ['base-hvac-install-quality-air-to-air-heat-pump-var-speed-detailed-performance.xml'].include? hpxml_file
-    # YORK HMH7
-    # https://ashp.neep.org/#!/product/64253/7/25000///0
-    clg_perf_data = hpxml_bldg.heat_pumps[0].cooling_detailed_performance_data
-    htg_perf_data = hpxml_bldg.heat_pumps[0].heating_detailed_performance_data
-    clg_perf_data.add(outdoor_temperature: 95.0,
-                      capacity: 11700,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 4.47)
-    clg_perf_data.add(outdoor_temperature: 95.0,
-                      capacity: 36000,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 2.71)
-    clg_perf_data.add(outdoor_temperature: 82.0,
-                      capacity: 13200,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 6.34)
-    clg_perf_data.add(outdoor_temperature: 82.0,
-                      capacity: 40000,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 3.53)
-    htg_perf_data.add(outdoor_temperature: 47.0,
-                      capacity: 10000,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 4.73)
-    htg_perf_data.add(outdoor_temperature: 47.0,
-                      capacity: 36000,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 3.44)
-    htg_perf_data.add(outdoor_temperature: 17.0,
-                      capacity: 4200,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 1.84)
-    htg_perf_data.add(outdoor_temperature: 17.0,
-                      capacity: 24800,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 2.66)
-    htg_perf_data.add(outdoor_temperature: 5.0,
-                      capacity: 1900,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 0.81)
-    htg_perf_data.add(outdoor_temperature: 5.0,
-                      capacity: 19900,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 2.28)
-    hpxml_bldg.heat_pumps[0].airflow_defect_ratio = -0.25
-    hpxml_bldg.heat_pumps[0].charge_defect_ratio = -0.25
-  end
-  if ['base-hvac-air-to-air-heat-pump-var-speed-detailed-performance-other-temperatures.xml'].include? hpxml_file
-    clg_perf_data = hpxml_bldg.heat_pumps[0].cooling_detailed_performance_data
-    htg_perf_data = hpxml_bldg.heat_pumps[0].heating_detailed_performance_data
-    clg_perf_data.add(outdoor_temperature: 95.0,
-                      capacity: 11700,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 4.47)
-    clg_perf_data.add(outdoor_temperature: 95.0,
-                      capacity: 36000,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 2.71)
-    clg_perf_data.add(outdoor_temperature: 105.0,
-                      capacity: 10000,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 3.9)
-    clg_perf_data.add(outdoor_temperature: 105.0,
-                      capacity: 30000,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 2.3)
-    htg_perf_data.add(outdoor_temperature: 47.0,
-                      capacity: 10000,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 4.73)
-    htg_perf_data.add(outdoor_temperature: 47.0,
-                      capacity: 36000,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 3.44)
-    htg_perf_data.add(outdoor_temperature: 55.0,
-                      capacity: 12000,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 5.5)
-    htg_perf_data.add(outdoor_temperature: 55.0,
-                      capacity: 45000,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 4.0)
-  end
-  if ['base-hvac-central-ac-only-var-speed-detailed-performance.xml'].include? hpxml_file
-    clg_perf_data = hpxml_bldg.cooling_systems[0].cooling_detailed_performance_data
-    clg_perf_data.add(outdoor_temperature: 95.0,
-                      capacity: 11700,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 4.47)
-    clg_perf_data.add(outdoor_temperature: 95.0,
-                      capacity: 36000,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 2.71)
-    clg_perf_data.add(outdoor_temperature: 82.0,
-                      capacity: 13200,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 6.34)
-    clg_perf_data.add(outdoor_temperature: 82.0,
-                      capacity: 40000,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 3.53)
-  end
-  if ['base-hvac-mini-split-air-conditioner-only-ductless-detailed-performance.xml'].include? hpxml_file
-    clg_perf_data = hpxml_bldg.cooling_systems[0].cooling_detailed_performance_data
-    clg_perf_data.add(outdoor_temperature: 95.0,
-                      capacity: 10372,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 4.05)
-    clg_perf_data.add(outdoor_temperature: 95.0,
-                      capacity: 42653,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 3.27)
-    clg_perf_data.add(outdoor_temperature: 82.0,
-                      capacity: 19456,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 8.03)
-    clg_perf_data.add(outdoor_temperature: 82.0,
-                      capacity: 40093,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 3.27)
-  end
-  if ['base-hvac-mini-split-heat-pump-ducted-detailed-performance.xml'].include? hpxml_file
-    # FUJITSU Halcyon Single-room Mini-Split Systems Slim
-    # https://ashp.neep.org/#!/product/25352/7/25000///0
-    clg_perf_data = hpxml_bldg.heat_pumps[0].cooling_detailed_performance_data
-    htg_perf_data = hpxml_bldg.heat_pumps[0].heating_detailed_performance_data
-    clg_perf_data.add(outdoor_temperature: 95.0,
-                      capacity: 9600,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 4.02)
-    clg_perf_data.add(outdoor_temperature: 95.0,
-                      capacity: 39000,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 2.86)
-    clg_perf_data.add(outdoor_temperature: 82.0,
-                      capacity: 10224,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 4.61)
-    clg_perf_data.add(outdoor_temperature: 82.0,
-                      capacity: 41587,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 3.29)
-    htg_perf_data.add(outdoor_temperature: 47.0,
-                      capacity: 9200,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 4.35)
-    htg_perf_data.add(outdoor_temperature: 47.0,
-                      capacity: 48000,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 3.21)
-    htg_perf_data.add(outdoor_temperature: 17.0,
-                      capacity: 7063,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 2.92)
-    htg_perf_data.add(outdoor_temperature: 17.0,
-                      capacity: 36800,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 2.15)
-    htg_perf_data.add(outdoor_temperature: 5.0,
-                      capacity: 6310,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 2.60)
-    htg_perf_data.add(outdoor_temperature: 5.0,
-                      capacity: 32920,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 1.93)
-  end
-  if ['base-hvac-mini-split-heat-pump-ductless-detailed-performance.xml'].include? hpxml_file
-    # BOSCH Bosch Climate 5000 ductless minisplit series
-    # https://ashp.neep.org/#!/product/66076/7/25000///0
-    clg_perf_data = hpxml_bldg.heat_pumps[0].cooling_detailed_performance_data
-    htg_perf_data = hpxml_bldg.heat_pumps[0].heating_detailed_performance_data
-    clg_perf_data.add(outdoor_temperature: 95.0,
-                      capacity: 10372,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 4.05)
-    clg_perf_data.add(outdoor_temperature: 95.0,
-                      capacity: 42653,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 3.27)
-    clg_perf_data.add(outdoor_temperature: 82.0,
-                      capacity: 19456,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 8.03)
-    clg_perf_data.add(outdoor_temperature: 82.0,
-                      capacity: 40093,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 3.27)
-    htg_perf_data.add(outdoor_temperature: 47.0,
-                      capacity: 12143,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 4.81)
-    htg_perf_data.add(outdoor_temperature: 47.0,
-                      capacity: 56499,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 3.17)
-    htg_perf_data.add(outdoor_temperature: 17.0,
-                      capacity: 7414,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 1.96)
-    htg_perf_data.add(outdoor_temperature: 17.0,
-                      capacity: 43387,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 2.31)
-    htg_perf_data.add(outdoor_temperature: 5.0,
-                      capacity: 8130,
-                      capacity_description: HPXML::CapacityDescriptionMinimum,
-                      efficiency_cop: 1.71)
-    htg_perf_data.add(outdoor_temperature: 5.0,
-                      capacity: 36037,
-                      capacity_description: HPXML::CapacityDescriptionMaximum,
-                      efficiency_cop: 1.96)
   end
 
   hpxml.buildings.each do |hpxml_bldg|
@@ -2585,10 +2327,16 @@ if ARGV[0].to_sym == :update_hpxmls
   ENV['HOMEDRIVE'] = 'C:\\' if !ENV['HOMEDRIVE'].nil? && ENV['HOMEDRIVE'].start_with?('U:')
 
   # Create sample/test HPXMLs
-  OpenStudio::Logger.instance.standardOutLogger.setLogLevel(OpenStudio::Fatal)
   t = Time.now
   create_hpxmls()
   puts "Completed in #{(Time.now - t).round(1)}s"
+
+  # Reformat real_homes HPXMLs
+  puts 'Reformatting real_homes HPXMLs...'
+  Dir['workflow/real_homes/*.xml'].each do |hpxml_path|
+    hpxml = HPXML.new(hpxml_path: hpxml_path)
+    XMLHelper.write_file(hpxml.to_doc, hpxml_path)
+  end
 end
 
 if ARGV[0].to_sym == :download_utility_rates